Move vite-tsconfig-paths to devDeps, useThemedFavicon hook
