+++ package.json
@@ -14,8 +14,7 @@
     "@reduxjs/toolkit": "^1.9.7",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.1.3",
-    "vite-tsconfig-paths": "^4.2.1"
+    "react-redux": "^8.1.3"
   },
   "devDependencies": {
     "@emotion/babel-plugin": "^11.11.0",
@@ -34,6 +33,7 @@
     "husky": "^8.0.3",
     "prettier": "2.8.0",
     "typescript": "^4.9.5",
-    "vite": "^3.2.7"
+    "vite": "^3.2.7",
+    "vite-tsconfig-paths": "^4.2.1"
   }
 }
