+++ src/components/Layout/index.tsx
@@ -1,20 +1,9 @@
-import { PropsWithChildren, useEffect } from 'react';
+import { PropsWithChildren } from 'react';
 import { Box } from 'components/Box';
-import { useAppSelector } from 'store';
-import { selectCurrentInterval } from 'store/intervals';
-import { selectTheme } from 'store/theme';
-
-const link = document.getElementById('favicon') as HTMLLinkElement;
+import { useThemedFavicon } from 'hooks';
 
 export const Layout = ({ children }: PropsWithChildren) => {
-  const currentInterval = useAppSelector(selectCurrentInterval);
-  const theme = useAppSelector(selectTheme);
-
-  useEffect(() => {
-    if (link) {
-      link.href = `/images/favicon/${currentInterval}-${theme}.svg`;
-    }
-  }, [currentInterval, theme]);
+  useThemedFavicon();
 
   return (
     <Box
